Added code for the next business questions:
Which actors most frequently appear in movies that are over 150 minutes long and belong to the "Drama" genre?

Which movies released after 2010 have a rating below 5 and received more than 5,000 votes?

How many documentary films were released each year between 1990 and 2024?

Which TV shows had the highest number of seasons and who was their main director?

Which actors made their debut in movies before the age of 25 and had the highest ratings for their debut roles?

For each genre, which movies are in the top 3 by number of votes?